% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc_stats.R
\name{biocStats}
\alias{biocStats}
\title{Get Bioconductor statistics}
\usage{
biocStats(pkg)
}
\arguments{
\item{pkg}{A Bioconductor package title.}
}
\description{
Pulls package data from Bioconductor and plots number of
   downloads from distinct IP addresses and total numbers, regardless of
   IP address distinction. Requires an active internet connection.
}
