% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brb.R
\name{brb}
\alias{brb}
\title{Be right back message prompt}
\usage{
brb(t, unit = c("m", "s", "h"))
}
\arguments{
\item{t}{Time}

\item{unit}{Unit of time `t`. Defaults to minute (`m`).}
}
\description{
Be right back message prompt
}
