% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{createDirs}
\alias{createDirs}
\title{Create Subdirectories}
\usage{
createDirs(p = NULL)
}
\arguments{
\item{p}{A character string specifying the path to the parent
directory where subdirectories should be created. If `NULL`,
the subdirectories will be created in the current working
directory. Default is `NULL`.}
}
\value{
The function creates directories as a side effect
and returns `NULL` invisibly.
}
\description{
This function creates a set of standard subdirectories ("data",
"src", "output", "notebook") in the current working directory
or in a specified parent directory.
}
\examples{
# Create directories in the current working directory
createDirs()

# Create directories inside an existing parent directory
createDirs("project_folder")

}
\seealso{
\code{\link{dir.create}}, \code{\link{file.path}}
}
