% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocStats.R
\name{getBiocStats}
\alias{getBiocStats}
\title{Get Bioconductor statistics}
\usage{
getBiocStats(pkg)
}
\arguments{
\item{pkg}{A Bioconductor package title.}
}
\description{
Pulls package data from Bioconductor and plots number of
   downloads from distinct IP addresses and total numbers, regardless of
   IP address distinction. Requires an active internet connection.
}
