% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{oneHot}
\alias{oneHot}
\title{One-hot encode a DNA sequence}
\usage{
oneHot(s, n = c("A", "C", "G", "T"))
}
\arguments{
\item{s}{A character string of a DNA sequence.}

\item{n}{A character vector of given nucleotide elements. Defaults to
\code{c("A", "C", "G", "T")}.}
}
\description{
Takes a DNA string and returns a one-hot encoded matrix of
   4 rows \code{c("A", "C", "G", "T")} with \eqn{j}-columns, where
   \eqn{j} is the number of characters in a DNA string.
}
