% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_bar.R
\name{propBar}
\alias{propBar}
\title{Make text-based proportion bars in the console}
\usage{
propBar(
  use,
  total,
  charUse = "#",
  charRem = " ",
  charLen = 25,
  charEnd = c("[", "]")
)
}
\arguments{
\item{use}{Number of units used in total size.}

\item{total}{Total number of units in total size.}

\item{charUse}{Text character to display used portion of bar. Defaults to
\code{"#"}.}

\item{charRem}{Text character to display unused portion of bar. Defaults to
\code{" "}.}

\item{charLen}{Total number of characters used for bar. Default is 25 units.}

\item{charEnd}{"Fancy" end pieces if you want them. Must be supplied as a
vector of character elements. Defaults to \code{c("[", "]")}.}
}
\description{
Generates a text-based proportion bar for downstream
   applications.
}
